ğ

+++ Brackey's Game Jam 2023.2/Assets/Scripts/PurpleBoss.cs
@@ -41,7 +41,7 @@ public class PurpleBoss : MonoBehaviour
             sfx.PlayEnemyHit();
             if (purpleBossHealth == 0)
             {
-                PlayerPrefs.SetFloat("PurpleTrigger", 0);
+                PlayerPrefs.SetInt("PurpleTrigger", 0);
                 StartCoroutine(StartCutScene());
                 
             }
